@@ -1,13 +1,14 @@
 plugins {
-    id 'org.springframework.boot' version '2.7.2'
-    id 'io.spring.dependency-management' version '1.0.12.RELEASE'
+    id 'org.springframework.boot' version '3.0.5'
+    id 'io.spring.dependency-management' version '1.1.0'
     id 'java'
     id 'jacoco'
+    id "org.sonarqube" version "4.0.0.2929"
 }
 
 group = 'crawler'
-version = '0.2.0'
-sourceCompatibility = '11'
+version = '1.0.0'
+sourceCompatibility = '17'
 
 configurations {
     developmentOnly
@@ -30,8 +31,8 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-webflux'
     implementation 'org.apache.lucene:lucene-analyzers-kuromoji:8.11.1'
-    implementation 'org.hibernate.search:hibernate-search-mapper-orm:6.1.6.Final'
-    implementation 'org.hibernate.search:hibernate-search-backend-lucene:6.1.6.Final'
+    implementation 'org.hibernate.search:hibernate-search-mapper-orm-orm6:6.1.8.Final'
+    implementation 'org.hibernate.search:hibernate-search-backend-lucene:6.1.8.Final'
     compileOnly 'io.jsonwebtoken:jjwt-api:0.11.5'
     compileOnly 'org.projectlombok:lombok'
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
