@@ -6,20 +6,21 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import javax.persistence.Basic;
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.Lob;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.persistence.Basic;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Lob;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 import org.hibernate.search.engine.backend.types.Sortable;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.FullTextField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexedEmbedded;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.KeywordField;
 
 import lombok.AllArgsConstructor;
@@ -44,6 +45,15 @@ import lombok.Setter;
 @XmlRootElement
 public class Novel extends BaseObject implements Serializable {
 
+    /** タイトルのフィールド名 */
+    public static final String TITLE_FIELD_NAME = "title";
+
+    /** 作者名のフィールド名 */
+    public static final String WRITERNAME_FIELD_NAME = "writername";
+
+    /** 解説のフィールド名 */
+    public static final String DESCRIPTION_FIELD_NAME = "description";
+
     /** URL */
     @Column(nullable = false, length = 64)
     private String url;
@@ -52,21 +62,21 @@ public class Novel extends BaseObject implements Serializable {
     @EqualsAndHashCode.Exclude
     @Column(length = 100)
     @FullTextField(analyzer = "japanese")
-    @KeywordField(name = "titleSort", sortable = Sortable.YES)
+    @KeywordField(name = TITLE_FIELD_NAME + SORT_NAME, sortable = Sortable.YES)
     private String title;
 
     /** 作者名 */
     @EqualsAndHashCode.Exclude
     @Column(length = 100)
     @FullTextField(analyzer = "japanese")
-    @KeywordField(name = "writernameSort", sortable = Sortable.YES)
+    @KeywordField(name = WRITERNAME_FIELD_NAME + SORT_NAME, sortable = Sortable.YES)
     private String writername;
 
     /** 解説 */
     @EqualsAndHashCode.Exclude
     @Column
     @FullTextField(analyzer = "japanese")
-    @KeywordField(name = "descriptionSort", sortable = Sortable.YES)
+    @KeywordField(name = DESCRIPTION_FIELD_NAME + SORT_NAME, sortable = Sortable.YES)
     private String description;
 
     /** 本文 */
@@ -84,6 +94,7 @@ public class Novel extends BaseObject implements Serializable {
     /** 小説の付随情報 */
     @EqualsAndHashCode.Exclude
     @OneToOne(fetch = FetchType.LAZY, mappedBy = "novel", cascade = CascadeType.ALL)
+    @IndexedEmbedded
     private NovelInfo novelInfo;
 
     /** 小説の更新履歴セット */
