@@ -3,24 +3,29 @@ package crawlerapi.entity;
 import java.io.Serializable;
 import java.time.Duration;
 import java.time.LocalDateTime;
-import java.util.HashSet;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
-import javax.persistence.Transient;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import jakarta.persistence.Transient;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.hibernate.search.engine.backend.types.Aggregable;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.FullTextField;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexingDependency;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.KeywordField;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.ObjectPath;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.PropertyValue;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -38,6 +43,7 @@ import lombok.Setter;
 @Builder
 @Entity
 @Table(name = "novel_info")
+@Indexed
 public class NovelInfo extends BaseObject implements Serializable {
 
     /** ログ出力クラス */
@@ -57,13 +63,8 @@ public class NovelInfo extends BaseObject implements Serializable {
 
     /** キーワード */
     @Column(length = 300)
+    @FullTextField(analyzer = "whitespace")
     private String keyword;
-
-    /** キーワードセット */
-    @Builder.Default
-    @Transient
-    @OneToMany
-    private Set<KeywordWrap> keywordSet = new HashSet<>();
 
     /** お気に入りフラグ */
     @Column
@@ -113,28 +114,15 @@ public class NovelInfo extends BaseObject implements Serializable {
     }
 
     /**
-     * キーワードを設定する.
-     * スペースで分割したキーワードをKeywordWrapに設定する.
+     * スペースで分割したキーワードを取得する.
      *
-     * @param keyword
-     *            キーワード
+     * @return スペースで分割したキーワード
      */
-    public void setKeyword(String keyword) {
-        this.keyword = keyword;
-
-        Stream.of(Optional.ofNullable(keyword).orElseGet(String::new).split(" "))
-                .collect(Collectors.toSet()).forEach(keywords -> keywordSet.add(new KeywordWrap(keywords)));
+    @Transient
+    @IndexingDependency(derivedFrom = @ObjectPath(@PropertyValue(propertyName = "keyword")))
+    @KeywordField(name = "keyword_facet", aggregable = Aggregable.YES)
+    public Set<String> getKeywordSet() {
+        return Stream.of(Optional.ofNullable(keyword).orElseGet(String::new).split(" "))
+                .collect(Collectors.toSet());
     }
-}
-
-/**
- * 小説の付随情報のキーワード
- */
-@AllArgsConstructor
-@Setter
-@Getter
-class KeywordWrap implements Serializable {
-
-    /** キーワード */
-    String keyword;
 }
