@@ -1,12 +1,13 @@
 package crawlerapi.repository;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.stereotype.Repository;
 
 import crawlerapi.entity.Novel;
 
-@Repository
 public interface NovelRepository extends JpaRepository<Novel, Long>, JpaSpecificationExecutor<Novel> {
 
+    List<Novel> findAllByOrderByTitleAscWriternameAsc();
 }
